ProductReleaseAnnouncement/requirements.txt
requests>=2.32
markdown>=3.7
docopt-ng>=0.9.0
